# Specialize arc store for continuous label in FST