--- a/lucene/core/src/java/org/apache/lucene/util/fst/FSTEnum.java
+++ b/lucene/core/src/java/org/apache/lucene/util/fst/FSTEnum.java
@@ -149,12 +149,57 @@ abstract class FSTEnum<T> {
         final FST.BytesReader in = fst.getBytesReader();
         if (arc.nodeFlags() == FST.ARCS_FOR_DIRECT_ADDRESSING) {
           arc = doSeekCeilArrayDirectAddressing(arc, targetLabel, in);
-        } else {
-          assert arc.nodeFlags() == FST.ARCS_FOR_BINARY_SEARCH;
+        } else if (arc.nodeFlags() == FST.ARCS_FOR_BINARY_SEARCH) {
           arc = doSeekCeilArrayPacked(arc, targetLabel, in);
+        } else {
+          assert arc.nodeFlags() == FST.ARCS_FOR_CONTINUOUS;
+          arc = doSeekCeilArrayContinuous(arc, targetLabel, in);
         }
       } else {
         arc = doSeekCeilList(arc, targetLabel);
+      }
+    }
+  }
+
+  private FST.Arc<T> doSeekCeilArrayContinuous(
+      final FST.Arc<T> arc, final int targetLabel, final FST.BytesReader in) throws IOException {
+    int targetIndex = targetLabel - arc.firstLabel();
+    if (targetIndex >= arc.numArcs()) {
+      // Target is beyond the last arc, out of label range.
+      // Dead end (target is after the last arc);
+      // rollback to last fork then push
+      upto--;
+      while (true) {
+        if (upto == 0) {
+          return null;
+        }
+        final FST.Arc<T> prevArc = getArc(upto);
+        // System.out.println("  rollback upto=" + upto + " arc.label=" + prevArc.label + "
+        // isLast?=" + prevArc.isLast());
+        if (!prevArc.isLast()) {
+          fst.readNextArc(prevArc, fstReader);
+          pushFirst();
+          return null;
+        }
+        upto--;
+      }
+    } else {
+      if (targetIndex < 0) {
+        fst.readArcByContinuous(arc, in, 0);
+        assert arc.label() > targetLabel;
+        pushFirst();
+        return null;
+      } else {
+        fst.readArcByContinuous(arc, in, targetIndex);
+        assert arc.label() == targetLabel;
+        // found -- copy pasta from below
+        output[upto] = fst.outputs.add(output[upto - 1], arc.output());
+        if (targetLabel == FST.END_LABEL) {
+          return null;
+        }
+        setCurrentLabel(arc.label());
+        incr();
+        return fst.readFirstTargetArc(arc, getArc(upto), fstReader);
       }
     }
   }
@@ -332,13 +377,43 @@ abstract class FSTEnum<T> {
         final FST.BytesReader in = fst.getBytesReader();
         if (arc.nodeFlags() == FST.ARCS_FOR_DIRECT_ADDRESSING) {
           arc = doSeekFloorArrayDirectAddressing(arc, targetLabel, in);
-        } else {
-          assert arc.nodeFlags() == FST.ARCS_FOR_BINARY_SEARCH;
+        } else if (arc.nodeFlags() == FST.ARCS_FOR_BINARY_SEARCH) {
           arc = doSeekFloorArrayPacked(arc, targetLabel, in);
+        } else {
+          assert arc.nodeFlags() == FST.ARCS_FOR_CONTINUOUS;
+          arc = doSeekFloorContinuous(arc, targetLabel, in);
         }
       } else {
         arc = doSeekFloorList(arc, targetLabel);
       }
+    }
+  }
+
+  private FST.Arc<T> doSeekFloorContinuous(FST.Arc<T> arc, int targetLabel, FST.BytesReader in)
+      throws IOException {
+    int targetIndex = targetLabel - arc.firstLabel();
+    if (targetIndex < 0) {
+      // Before first arc.
+      return backtrackToFloorArc(arc, targetLabel, in);
+    } else if (targetIndex >= arc.numArcs()) {
+      // After last arc.
+      fst.readLastArcByContinuous(arc, in);
+      assert arc.label() < targetLabel;
+      assert arc.isLast();
+      pushLast();
+      return null;
+    } else {
+      // Within label range.
+      fst.readArcByContinuous(arc, in, targetIndex);
+      assert arc.label() == targetLabel;
+      // found -- copy pasta from below
+      output[upto] = fst.outputs.add(output[upto - 1], arc.output());
+      if (targetLabel == FST.END_LABEL) {
+        return null;
+      }
+      setCurrentLabel(arc.label());
+      incr();
+      return fst.readFirstTargetArc(arc, getArc(upto), fstReader);
     }
   }
 
@@ -400,9 +475,11 @@ abstract class FSTEnum<T> {
           if (arc.bytesPerArc() != 0 && arc.label() != FST.END_LABEL) {
             if (arc.nodeFlags() == FST.ARCS_FOR_BINARY_SEARCH) {
               findNextFloorArcBinarySearch(arc, targetLabel, in);
-            } else {
-              assert arc.nodeFlags() == FST.ARCS_FOR_DIRECT_ADDRESSING;
+            } else if (arc.nodeFlags() == FST.ARCS_FOR_DIRECT_ADDRESSING) {
               findNextFloorArcDirectAddressing(arc, targetLabel, in);
+            } else {
+              assert arc.nodeFlags() == FST.ARCS_FOR_CONTINUOUS;
+              findNextFloorArcContinuous(arc, targetLabel, in);
             }
           } else {
             while (!arc.isLast() && fst.readNextArcLabel(arc, in) < targetLabel) {
@@ -448,6 +525,24 @@ abstract class FSTEnum<T> {
         if (floorIndex > 0) {
           fst.readArcByDirectAddressing(arc, in, floorIndex);
         }
+      }
+    }
+  }
+
+  /** Same as {@link #findNextFloorArcDirectAddressing} for continuous node. */
+  private void findNextFloorArcContinuous(FST.Arc<T> arc, int targetLabel, final FST.BytesReader in)
+      throws IOException {
+    assert arc.nodeFlags() == FST.ARCS_FOR_CONTINUOUS;
+    assert arc.label() != FST.END_LABEL;
+    assert arc.label() == arc.firstLabel();
+    if (arc.numArcs() > 1) {
+      int targetIndex = targetLabel - arc.firstLabel();
+      assert targetIndex >= 0;
+      if (targetIndex >= arc.numArcs()) {
+        // Beyond last arc. Take last arc.
+        fst.readLastArcByContinuous(arc, in);
+      } else {
+        fst.readArcByContinuous(arc, in, targetIndex - 1);
       }
     }
   }