--- a/lucene/core/src/java/org/apache/lucene/util/fst/NodeHash.java
+++ b/lucene/core/src/java/org/apache/lucene/util/fst/NodeHash.java
@@ -409,7 +409,8 @@ final class NodeHash<T> {
             }
             break;
           case FST.ARCS_FOR_CONTINUOUS:
-            if ((node.arcs[node.numArcs - 1].label - node.arcs[0].label + 1) != scratchArc.numArcs()) {
+            if ((node.arcs[node.numArcs - 1].label - node.arcs[0].label + 1)
+                != scratchArc.numArcs()) {
               return -1;
             }
             break;