… lucene/core/src/test/org/apache/lucene/util/fst/TestUtil.java