This PR resolves issue: https://github.com/apache/lucene/issues/12701 . Thanks for the cool idea from @gf2121 

It need some more benchmarking.