@mikemccand 2023-11-02
This looks cool!  Sorry, I caused conflicts w/ the earlier merge -- could you please resolve those @easyice?  I'm happy to try benchmarking it, using the new `IndexToFST` tool in luceneutil :)