@mikemccand 2023-11-03
I tested this PR using `IndexToFST` from `luceneutil`.  This just tests construction time and final FST size, on all `wikimediumall` unique terms, allowing up to 64 MB RAM while building the FST:

`main`:

```
  saved FST to "fst.bin": 382070800 bytes; 44.146 sec
  saved FST to "fst.bin": 382070800 bytes; 43.478 sec
```


PR:

```
  saved FST to "fst.bin": 381705016 bytes; 42.616 sec
  saved FST to "fst.bin": 381705016 bytes; 42.832 sec
```

FST size is a wee bit smaller (~0.1%), and curiously the construction time seems to be faster too.