@mikemccand 2023-11-03
I'll run `Test2BFST` too ... takes a few hours!