@mikemccand 2023-11-03
`Test2BFSTs` is happy:

```
BUILD SUCCESSFUL in 50m 6s
```