@mikemccand 2023-11-06
This might be a nice bump in `PKLookup` performance in the nightly benchmarks -- it uses compact integers encoded as `BytesRef` in the `id` field.