@mikemccand 2023-11-09
> I can help merge this in and backport if there is no objection in 48h.

Thanks @gf2121 -- we should backport all these recent exciting FST changes in the right order as a batch to avoid scary hairy conflicts.  I plan to do this in the next few days -- they have been baking in `main` quite well recently.