@jpountz 2023-11-20
@mikemccand is now a good time to backport these changes?